requirements.txt
asgiref==3.7.2
Django==5.0.1
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
pillow==10.2.0
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
setuptools==69.0.3
sqlparse==0.4.4
uritemplate==4.1.1